Add support for redirector and web proxy
Now, the custom bootstrap script talks to redirector to figure out
what cluster the tenant is hosted on. It uses one of the received
URLs to further talk to CVaaS.

This change also adds support for running the script (and hence, ZTP)
in an environment where internet (i.e. CVaaS here) can only be
accessed using a non-transparent web proxy.