Add support for redirector and web proxy (#2)
Now, the custom bootstrap script talks to redirector to figure out what cluster the tenant is hosted on. It uses one of the received URLs to further talk to CVaaS.

This change also adds support for running the script (and hence, ZTP) in an environment where internet (i.e. CVaaS here) can only be accessed using a non-transparent web proxy.

We have also added support for customers to do device upgrades if they want. This also improves and fixes a number of bugs in the code.